requirements.txt
Flask==0.12.2
Flask-PyMongo==0.3.0
Flask-RESTful==0.3.6
gunicorn==19.7.1
virtualenv==15.1.0
click==6.7
Jinja2==2.8
itsdangerous==0.24
MarkupSafe==0.23
nose==1.3.7
heroku==0.1.4
coverage==4.4.1
codecov==2.0.9
pymongo==2.8
pyjwt==1.4.2
mock==2.0.0
